Add: module init code & exception case
