+++ loadenv.js
@@ -55,6 +55,6 @@ module.exports = {
       endAt: "2024-03-19T00:00:00+09:00",
     },
   }, // optional
-  naverCloudApiId: process.env.NAVER_MAP_API_ID, // optional
-  naverCloudApiKey: process.env.NAVER_MAP_API_KEY, //optional
+  naverCloudApiId: process.env.NAVER_MAP_API_ID || "none", // optional
+  naverCloudApiKey: process.env.NAVER_MAP_API_KEY || "none", //optional
 };
