+++ src/services/fare.js
@@ -13,124 +13,6 @@ const naverCloudApi = {
 const naverCloudApiCall =
   "https://naveropenapi.apigw.ntruss.com/map-direction/v1/driving?start=";
 
-/**  Initialize database
- * 1. Erase all previous data
- * 2. Sets all taxi fare to 0
- * @summary 카이스트 본원 <-> 대전역의 경우 48 * 7개의 시간대에 대한 택시 요금을 init 시점의 비용으로 설정합니다.
- * @summary 카이스트 본원 <-> 대전역 외 나머지 경로, 238개의 경로에 대해서는 한 개의 collection씩 설정하여 fare를 init 시점의 비용으로 설정합니다. time은 0으로 설정합니다.
- */
-const initDatabase = async (req, res) => {
-  try {
-    // Remove all previous data
-    await taxiFareModel.deleteMany({});
-
-    const location = await locationModel.find({ isValid: { $eq: true } });
-
-    location.map((from) => {
-      location.map(async (to) => {
-        if (from._id === to._id) return;
-        let tableFare = [];
-        // 카이스트 본원 <-> 대전역의 경우 48*7(=336)개의 시간대에 대한 택시 요금을 일괄적으로 설정
-        if (from.koName === "카이스트 본원" && to.koName === "대전역") {
-          const fare = (
-            await axios.get(
-              `${
-                naverCloudApiCall + from.longitude + "," + from.latitude
-              }&goal=${to.longitude + "," + to.latitude}&options=traoptimal`,
-              { headers: naverCloudApi }
-            )
-          ).data.route.traoptimal[0].summary.taxiFare;
-          [...Array(48 * 7)].map((_, i) => {
-            tableFare.push({
-              from: from._id,
-              to: to._id,
-              time: i,
-              fare: fare,
-              isMajor: true,
-            });
-          });
-        } else if (from.koName === "대전역" && to.koName === "카이스트 본원") {
-          await axios
-            .get(
-              `${
-                naverCloudApiCall + from.longitude + "," + from.latitude
-              }&goal=${to.longitude + "," + to.latitude}&options=traoptimal`,
-              { headers: naverCloudApi }
-            )
-            .then((res) => {
-              [...Array(48 * 7)].map((_, i) => {
-                tableFare.push({
-                  from: from._id,
-                  to: to._id,
-                  time: i,
-                  fare: res.data.route.traoptimal[0].summary.taxiFare,
-                  isMajor: true,
-                });
-              });
-            })
-            .catch((err) => {
-              logger.error(err.message);
-              [...Array(48 * 7)].map((_, i) => {
-                tableFare.push({
-                  from: from._id,
-                  to: to._id,
-                  time: i,
-                  fare: 0,
-                  isMajor: true,
-                });
-              });
-            });
-        }
-        // 카이스트 본원 <-> 대전역외의 경로(238개)에 대해서는 7개(일주일) 씩 collection 지정 설정
-        else {
-          await new Promise(() =>
-            setTimeout(
-              async () =>
-                await axios
-                  .get(
-                    `${
-                      naverCloudApiCall + from.longitude + "," + from.latitude
-                    }&goal=${
-                      to.longitude + "," + to.latitude
-                    }&options=traoptimal`,
-                    { headers: naverCloudApi }
-                  )
-                  .then((res) => {
-                    [...Array(7)].map((_, i) => {
-                      tableFare.push({
-                        from: from,
-                        to: to,
-                        time: i * 48,
-                        fare: res.data.route.traoptimal[0].summary.taxiFare,
-                        isMajor: false,
-                      });
-                    });
-                  })
-                  .catch((err) => {
-                    logger.error(err.message);
-                    [...Array(7)].map((_, i) => {
-                      tableFare.push({
-                        from: from,
-                        to: to,
-                        time: i * 48,
-                        fare: 0,
-                        isMajor: false,
-                      });
-                    });
-                  }),
-              100
-            )
-          );
-        }
-        await taxiFareModel.insertMany(tableFare);
-      });
-    });
-    res.status(200).send("TaxiFare Database initialized");
-  } catch (err) {
-    res.status(500).json({ error: "fare/init: TaxiFare Database failed" });
-  }
-};
-
 /**
  * 주어진 from, to, time에 대한 택시 요금을 반환합니다.
  * @summary 카이스트 본원 <-> 대전역의 경로를 제외한 다른 경로의 경우, 1주일 전 매일 18:00시의 택시 요금을 반환합니다.
@@ -142,6 +24,18 @@ const initDatabase = async (req, res) => {
  */
 const getTaxiFare = async (req, res) => {
   try {
+    if (
+      naverCloudApi["X-NCP-APIGW-API-KEY"] == "none" ||
+      naverCloudApi["X-NCP-APIGW-API-KEY-ID"] == "none"
+    ) {
+      logger.log(
+        "Naver Cloud API가 존재하지 않습니다.택시 비용 관련 기능을 사용할 수 없습니다."
+      );
+      res
+        .status(503)
+        .json({ error: "fare/getTaxiFare: Naver Cloud API not found" });
+      return;
+    }
     const from = await locationModel
       .findOne({
         _id: { $eq: req.query.from },
@@ -300,7 +194,6 @@ const updateTaxiFare = async (sTime, isMajor) => {
 };
 
 module.exports = {
-  initDatabase,
   getTaxiFare,
   updateTaxiFare,
 };
